Add files via upload
Leveraging the power of APIs that is the last project in CodeCademy Introduction to JavaScript course. This website uses super cool Foursquare API to get data about points of interest in millions of locations around the globe and then Apixu API to get the weather forecast for this location. To tie it all together I used async...await and worked with JSON.